[IMCC] Fix builds on amd64. r32694 changed some variables to size_t, but now a pointer to a size_t was passed to get_codesize() as an int*.  On linux/amd64, int is 32 bit and size_t is 64 bit, thus, the high 32 bits were uninitialized and miniparrot crashed immediately when trying to allocate a 140TB buffer.
git-svn-id: https://svn.parrot.org/parrot/trunk@32716 d31e2699-5ff4-0310-a27c-f18f2fbe73fe

+++ compilers/imcc/pbc.c
@@ -129,9 +129,9 @@ static void fixup_globals(PARROT_INTERP)
         __attribute__nonnull__(1);
 
 PARROT_WARN_UNUSED_RESULT
-static int get_codesize(PARROT_INTERP,
+static size_t get_codesize(PARROT_INTERP,
     ARGIN(const IMC_Unit *unit),
-    ARGOUT(int *src_lines))
+    ARGOUT(size_t *src_lines))
         __attribute__nonnull__(1)
         __attribute__nonnull__(2)
         __attribute__nonnull__(3)
@@ -592,11 +592,11 @@ Stores globals for later fixup, returning the code size in number of ops.
 */
 
 PARROT_WARN_UNUSED_RESULT
-static int
-get_codesize(PARROT_INTERP, ARGIN(const IMC_Unit *unit), ARGOUT(int *src_lines))
+static size_t
+get_codesize(PARROT_INTERP, ARGIN(const IMC_Unit *unit), ARGOUT(size_t *src_lines))
 {
     Instruction *ins = unit->instructions;
-    int          code_size;
+    size_t       code_size;
 
     /* run through instructions:
      * - sanity check
